+++ ell-studio/src/components/evaluations/EvaluationCardTitle.js
@@ -1,6 +1,7 @@
 import React from "react";
-import { FiBarChart2, FiZap } from "react-icons/fi";
+import { FiZap } from "react-icons/fi";
 import VersionBadge from "../VersionBadge";
+import EvaluationsIcon from "./EvaluationsIcon";
 
 export function EvaluationCardTitle({
     evaluation,
@@ -30,8 +31,8 @@ export function EvaluationCardTitle({
 
     return (
         <div className={`flex items-center space-x-2 ${paddingClass} ${scaleClass} transition-colors ${additionalClassName} ${hoverClass} ${cursorClass} rounded-md overflow-hidden`} {...(clickable ? rest : {})}>
-            <div className="flex-shrink-0">
-                <FiBarChart2 className="h-4 w-4 text-yellow-600" />
+            <div className="flex-shrink-0 mr-1">
+                <EvaluationsIcon className="w-4 h-4 text-yellow-600" />
             </div>
             {nameOverride ? nameOverride : (
                 <code className={`px-2 py-1 rounded-md ${outlineClasses[outlineStyle]} text-${fontSize} font-medium truncate`}>
